# src/main/java/com/rabobank/csp/model/CustomerStatementProcessorRequest.java
package com.rabobank.csp.model;

import java.util.ArrayList;
import java.util.Objects;


/**
 * Data class for Customer statement processor request.
 */
public class CustomerStatementProcessorRequest extends ArrayList<CustomerStatementItem> {

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return Objects.hash(super.hashCode());
    }


    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append("class CustomerStatementProcessorRequestDTO {\n");
        sb.append("    ").append(toIndentedString(super.toString())).append("\n");
        sb.append("}");
        return sb.toString();
    }

    /**
     * Convert the given object to string with each line indented by 4 spaces
     * (except the first line).
     */
    private String toIndentedString(Object o) {
        if (o == null) {
            return "null";
        }
        return o.toString().replace("\n", "\n    ");
    }

}